allow enabling `preventEmojiSpam` on some channels (#171)
* ‌

* add roles bypass

* fix

* ‌

+++ src/features/preventEmojiSpam/index.ts
@@ -3,6 +3,8 @@ import { Events } from 'discord.js'
 import { definePlugin } from '@/types/definePlugin'
 import isOnlyEmoji from '@/utils/isOnlyEmoji'
 
+import { isEmojiPreventionEnabled } from './isEmojiPreventionEnabled'
+
 export default definePlugin({
   name: 'preventEmojiSpam',
   setup: (pluginContext) => {
@@ -11,13 +13,16 @@ export default definePlugin({
       once: false,
       execute: async (botContext, message) => {
         const { runtimeConfiguration, log } = botContext
+        const config = runtimeConfiguration.data.preventEmojiSpam
         // if has only emoji -> delete message
-        if (isOnlyEmoji(message.content)) {
+        if (isOnlyEmoji(message.content) && message.member) {
           try {
             log.info(
-              `Message with only emoji: ${message.id} by ${message.author}`,
+              `Message with only emoji: ${message.id} by ${message.member}`,
             )
-            if (runtimeConfiguration.data.preventEmojiSpam.enabled) {
+            if (
+              isEmojiPreventionEnabled(config, message.channel, message.member)
+            ) {
               await message.delete()
             }
           } catch (error) {
